+++ client/src/about/components/About.tsx
@@ -1,22 +1,8 @@
 import React, { FC } from 'react';
 import './About.scss';
-import { sendUpdateMessage, sendLoginMessage } from '../../websocket/actions'
 
-
-
-
-interface Props {
-  sendUpdateMessage: (id, x, y, r, g, b) => void;
-  sendLoginMessage: (id) => void;
-}
-
-const About: FC<Props> = ({ sendUpdateMessage, sendLoginMessage }) => (
+const About: FC = () => (
   <div className='about-page'>
-    <h2>TESTING AREA</h2>
-      <p>Click "Pixel 1" to send an update message to the server!
-      </p>
-        <button onClick = {() => sendUpdateMessage("user1", 10, 400, 255, 255, 255)} id="p1"> Pixel 1 </button>
-        <button onClick = {() => sendLoginMessage("testemail@gmail.com")} id="login"> User Login </button>
     <h1>OwlPlace</h1>
 
     <h1>About</h1>
